contrib/netsurf/libhubbub/test/data/tree-construction/after-body.dat

#data
<!DOCTYPE html><body></body>

#errors
#comments
Space characters in "after body" should be appended to the body node.
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "
"

#data
<!DOCTYPE html><body></body><!--x-->
#errors
#comments
Comment tokens in "after body" should be appended to the "html" element.
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|   <!-- x -->

#data
<!DOCTYPE html><body></body><!DOCTYPE html>
#errors
Unexpected DOCTYPE declaration.
#comments
DOCTYPE tokens in "after body" should be ignored.
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE html><body></body><html>
#errors
Second "html" element not allowed.
#comments
"html" elements in "after body" should be handled like in body and ignored.
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE html><body></body><html bgcolor="red">
#errors
Second "html" element not allowed.
#comments
"html" elements in "after body" should be handled like in body and ignored,
but extra attributes should be added to the root "html" element.
#document
| <!DOCTYPE html>
| <html>
|   bgcolor="red"
|   <head>
|   <body>

#data
<!DOCTYPE html><body></body></html>
#errors
#comments
End "html" tag should not result in parse error.
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE html><body></body>
#errors
#comments
EOF should not result in parse error.
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE html><body></body>xx
#errors
Content after end "body" tag.
#comments
Non-whitespace character tokens should raise parse error, and be processed
as "in body".
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "xx"

#data
<!DOCTYPE html><body></body><p>
#errors
Start tag after end "body" tag.
#comments
Non-html start tags should switch insertion mode to "in body" and processed
appropriately.
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>

#data
<!DOCTYPE html><body></body></dd>
#errors
End tag after end "body" tag.
End tag does not match any opened elements.
#comments
Non-html end tags should switch insertion mode to "in body" and processed
appropriately.
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>

#data
